+++ Modules/Intent.Modules.AspNetCore.Cors/release-notes.md
@@ -1,3 +1,8 @@
+### Version 3.3.11
+
+- Changed module name from `Intent.Modules.AspNetCore.Cors` to `Intent.AspNetCore.Cors`.
+- Changed to be configuration driven as opposed to hard coded and allowing all origins.
+
 ### Version 3.3.10
 
 - Updated supported client version to [3.3.0-pre.0, 5.0.0-a).